+++ Source/NexusLink/Private/Capabilities/Asset/WorldPartition/NexusCreateAssetDataLayerCapability.cpp
@@ -102,6 +102,11 @@ REGISTER_MCP_CAPABILITY(FCreateAssetDataLayerCapability)
 
 #else // NX_UE_HAS_DATA_LAYER_ASSET
 
+#include "Utils/NexusCapabilityResultBuilder.h"
+#include "NexusCapabilityRegistry.h"
+#include "NexusMcpSchemaBuilder.h"
+#include "NexusMcpTool.h"
+
 void FCreateAssetDataLayerCapability::BuildDefinition(FNexusCapabilityDefinition& Out) const
 {
 	Out.Name        = TEXT("create_asset_data_layer");